requirements.txt
# Essential Package
django
djangorestframework
django-cors-headers
mysqlclient

# Improve code quality Tool
black
isort
